requirements.txt
duckdb
PyYAML
numpy
pandas